TrueCrypt Source Version 4.3

+++ Linux/build.sh
@@ -1,8 +1,15 @@
 #!/bin/sh 
-# TrueCrypt build script
+#
+# Copyright (c) TrueCrypt Foundation. All rights reserved.
+#
+# Covered by the TrueCrypt License 2.2 the full text of which is contained
+# in the file License.txt included in TrueCrypt binary and source code
+# distribution packages.
+#
+
+KERNEL_VER=$(uname -r)
 
 TMP=.build.sh.tmp
-
 umask 022
 
 error ()
@@ -19,7 +26,7 @@ check_kernel_version ()
 	VER=$VER.$(grep '^PATCHLEVEL *=' "$M" | head -n 1 | tr -d ' ' | cut -d'=' -f2)
 	VER=$VER.$(grep '^SUBLEVEL *=' "$M" | head -n 1 | tr -d ' ' | cut -d'=' -f2)
 
-	[ $VER = $(uname -r | cut -d- -f1 | cut -d. -f1-3) ] && return 0
+	[ $VER = $(echo $KERNEL_VER | cut -d- -f1 | cut -d. -f1-3) ] && return 0
 	return 1
 }
 
@@ -30,7 +37,7 @@ echo "Checking build requirements..."
 [ $(id -u) -ne 0 ] && error "Administrator (root) privileges required for kernel source configuration." && exit 1
 
 V=""
-case "$(uname -r)" in
+case "$KERNEL_VER" in
 	[01].*) V=1 ;;
 	2.[0-5].*) V=1 ;;
 	2.6.[0-4]) V=1 ;;
@@ -38,15 +45,16 @@ case "$(uname -r)" in
 esac
 [ "$V" ] && error "TrueCrypt requires Linux kernel 2.6.5 or later" && exit 1
 
-KERNEL_SRC=/usr/src/linux-source-$(uname -r)
-check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-source-$(uname -r | cut -d'-' -f1)
-check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-$(uname -r)
-check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-$(uname -r | cut -d'-' -f1)
+KERNEL_SRC=/usr/src/linux-source-$KERNEL_VER
+check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-source-$(echo $KERNEL_VER | cut -d'-' -f1)
+check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-$KERNEL_VER
+check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux-$(echo $KERNEL_VER | cut -d'-' -f1)
+check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/kernels/$KERNEL_VER-$(uname -p)
 check_kernel_version "$KERNEL_SRC" || KERNEL_SRC=/usr/src/linux
 
 if ! check_kernel_version "$KERNEL_SRC"
 then
-	echo -n "Linux kernel ($(uname -r)) source directory [$KERNEL_SRC]: "
+	echo -n "Linux kernel ($KERNEL_VER) source directory [$KERNEL_SRC]: "
 	read A
 	[ "$A" ] && KERNEL_SRC="$A"
 	[ ! -d "$KERNEL_SRC" ] && error "$KERNEL_SRC does not exit" && exit 1
@@ -54,21 +62,21 @@ fi
 
 if ! check_kernel_version "$KERNEL_SRC"
 then
-	error "Kernel source version in $KERNEL_SRC is not $(uname -r)"
+	error "Kernel source version in $KERNEL_SRC is not $KERNEL_VER"
 	exit 1
 fi
 
 if [ ! -f "$KERNEL_SRC/drivers/md/dm.h" ]
 then
-	error "Kernel source code is incomplete - drivers/md/dm.h not found."
+	error "Kernel source code is incomplete - $KERNEL_SRC/drivers/md/dm.h not found."
 	exit 1
 fi
 
 if [ ! -f "$KERNEL_SRC/.config" ]
 then
-	if [ -f /proc/config.gz -o -f /boot/config-$(uname -r) ]
+	if [ -f /proc/config.gz -o -f /boot/config-$KERNEL_VER -o -f /boot/config-$(uname -r) ]
 	then
-		echo -n "Configure kernel source according to the currently running kernel? [Y/n]: "
+		echo -n "Configure kernel source according to the system configuration? [Y/n]: "
 		read A
 		if [ -z "$A" -o "$A" = "y" -o "$A" = "Y" ]
 		then
@@ -78,10 +86,15 @@ then
 			then
 				zcat /proc/config.gz >$KERNEL_SRC/.config || exit 1
 			else
-				cp /boot/config-$(uname -r) $KERNEL_SRC/.config || exit 1
+				if [ -f /boot/config-$(uname -r) ]
+				then
+					cp /boot/config-$(uname -r) $KERNEL_SRC/.config || exit 1
+				else
+					cp /boot/config-$KERNEL_VER $KERNEL_SRC/.config || exit 1
+				fi
 			fi
 			
-			make -C $KERNEL_SRC oldconfig </dev/zero >/dev/null || exit 1
+			make -C $KERNEL_SRC oldconfig </dev/null >/dev/null || exit 1
 			echo Done.
 		fi
 	fi
@@ -93,36 +106,36 @@ then
 	fi
 fi
 
-if [ ! -f "$KERNEL_SRC/scripts/modpost" -a ! -f "$KERNEL_SRC/scripts/mod/modpost" ]
+if [ ! -d "$KERNEL_SRC/include/asm" ] && grep -q modules_prepare $KERNEL_SRC/Makefile
 then
-	if grep -q modules_prepare $KERNEL_SRC/Makefile
+	echo -n "Preparing kernel build system in $KERNEL_SRC... "
+	if ! make -C $KERNEL_SRC modules_prepare >/dev/null 2>$TMP
 	then
-		echo -n "Preparing kernel build system in $KERNEL_SRC... "
-		if ! make -C $KERNEL_SRC modules_prepare >/dev/null 2>$TMP
-		then
-			cat $TMP; rm $TMP
-			exit 1
-		fi
-		rm $TMP
-		echo Done.
-	else
-		error "Kernel build system not ready. You should run make -C $KERNEL_SRC modules"
+		cat $TMP; rm $TMP
 		exit 1
 	fi
+	rm $TMP
+	echo Done.
 fi
 
 grep -qi 'CONFIG_BLK_DEV_DM=[YM]' $KERNEL_SRC/.config || echo "Warning: kernel device mapper support (CONFIG_BLK_DEV_DM) is disabled in $KERNEL_SRC"
 
-if [ ! -f "$KERNEL_SRC/Module.symvers" ] && grep -qi 'CONFIG_MODVERSIONS=Y' $KERNEL_SRC/.config 
+if [ ! -d "$KERNEL_SRC/include/asm" -o ! -f "$KERNEL_SRC/Module.symvers" ] 
 then
-		echo -n "Building internal kernel modules (may take a long time)... "
-		if ! make -C $KERNEL_SRC modules >/dev/null 2>$TMP
-		then
-			cat $TMP; rm $TMP
-			exit 1
-		fi
-		rm $TMP
-		echo Done.
+	echo -n "Building internal kernel modules (may take a long time)... "
+	if ! make -C $KERNEL_SRC modules >/dev/null 2>$TMP
+	then
+		cat $TMP; rm $TMP
+		exit 1
+	fi
+	rm $TMP
+	echo Done.
+fi
+
+if [ ! -d "$KERNEL_SRC/include/asm" ]
+then
+	error "Kernel source code is not prepared for building of modules - $KERNEL_SRC/include/asm not found."
+	exit 1
 fi
